refactor(configcore): PR-A5b cell.go split + errcode Category + ctx.Canceled classification
Main 1 — cell.go split (aligns with PR-A5a accesscore pattern):
- cell.go: package doc + compile-time checks + Option funcs + ConfigCore struct + NewConfigCore
- cell_init.go: Init + deriveModes + configDirectPublishMode + ensureCursorCodec + 6×init* methods
- cell_routes.go: RegisterRoutes + RegisterSubscriptions (with ref: k8s/kratos comments)

Main 2 — S15 errcode Category fix (config_repo.go):
- All ErrConfigRepoNotFound literal errors: +Category: errcode.CategoryDomain
- All ErrConfigRepoQuery literal errors: +Category: errcode.CategoryInfra
- errcode.Wrap(ErrConfigRepoQuery, ...) → errcode.WrapInfra (Create, List×4, PublishVersion)
- errcode.Wrap(ErrConfigDecryptFailed, ...) → errcode.WrapInfra (decryptValue, decryptVersionValue)
- ctx.Canceled / ctx.DeadlineExceeded short-circuit guard added to:
  scanConfigOrMapError (+ ctx parameter added), Update SELECT-FOR-UPDATE, GetVersion scan

Main 3 — TDD tests (config_repo_test.go):
- TestGetByKey_NotFound_HasDomainCategory: IsDomainNotFound + \!IsInfraError assertion
- TestConfigRepo_CtxCanceled_ClassifiedAsInfra: table-driven (Canceled + DeadlineExceeded)
  × 3 sites (GetByKey, Update_SelectForUpdate, GetVersion)
- Added IsDomainNotFound assertion to existing not-found tests:
  Update_NotFound, UpdateForRollback_NotFound, Delete_NotFound,
  GetVersion_NotFound_ReturnsErrConfigRepoNotFound (regression guard, 3c)

ref PR-A5a: accesscore cell.go split (cell_init.go + cell_routes.go pattern)
V-A15: dual-channel triage — infra errors must never map to domain not-found
