+++ cells/configcore/internal/adapters/postgres/config_repo.go
@@ -19,6 +19,17 @@ import (
 	"github.com/jackc/pgx/v5"
 )
 
+// ctxCanceledInfraError returns an infra error for context cancellation /
+// deadline exceeded conditions. Used as a pre-check in scan-or-map helpers to
+// prevent ctx.Canceled from being misclassified as a domain not-found event.
+func ctxCanceledInfraError(ctx context.Context, err error, op, key string) *errcode.Error {
+	if errors.Is(err, context.Canceled) || errors.Is(err, context.DeadlineExceeded) {
+		return errcode.WrapInfra(errcode.ErrConfigRepoQuery,
+			fmt.Sprintf("config repo: %s ctx canceled key=%s", op, key), err)
+	}
+	return nil
+}
+
 // DBTX abstracts the database operations needed by ConfigRepository.
 // Both pgxpool.Pool and pgx.Tx satisfy this interface.
 type DBTX interface {
@@ -148,7 +159,7 @@ func (r *ConfigRepository) decryptValue(ctx context.Context, key string, ct []by
 	aad := configcrypto.AADForConfig(cellID, key)
 	pt, err := r.transformer.Decrypt(ctx, ct, keyID, nonce, edk, aad)
 	if err != nil {
-		return "", errcode.Wrap(errcode.ErrConfigDecryptFailed,
+		return "", errcode.WrapInfra(errcode.ErrConfigDecryptFailed,
 			fmt.Sprintf("config repo: decrypt failed for key %s", key), err)
 	}
 	return string(pt), nil
@@ -182,7 +193,7 @@ func (r *ConfigRepository) decryptVersionValue(ctx context.Context, configID str
 	aad := configcrypto.AADForVersion(cellID, configID)
 	pt, err := r.transformer.Decrypt(ctx, ct, keyID, nonce, edk, aad)
 	if err != nil {
-		return "", errcode.Wrap(errcode.ErrConfigDecryptFailed,
+		return "", errcode.WrapInfra(errcode.ErrConfigDecryptFailed,
 			fmt.Sprintf("config repo: decrypt version failed for config_id %s", configID), err)
 	}
 	return string(pt), nil
@@ -231,7 +242,7 @@ func (r *ConfigRepository) Create(ctx context.Context, entry *domain.ConfigEntry
 		)
 	}
 	if err != nil {
-		return errcode.Wrap(errcode.ErrConfigRepoQuery,
+		return errcode.WrapInfra(errcode.ErrConfigRepoQuery,
 			fmt.Sprintf("config repo: create failed for key %s", entry.Key), err)
 	}
 	return nil
@@ -266,15 +277,23 @@ func scanConfigRow(row Row) (e *domain.ConfigEntry, valueCipher []byte, valueKey
 // scanConfigOrMapError calls scanConfigRow and translates the two known failure
 // modes into GoCell errcode.Error values:
 //
+//   - context.Canceled / context.DeadlineExceeded → ErrConfigRepoQuery (infra).
 //   - pgx.ErrNoRows  → ErrConfigRepoNotFound (domain not-found; maps to 404).
 //   - anything else  → ErrConfigRepoQuery (infra failure; maps to 500).
 //
+// The ctx.Canceled guard is checked first (before pgx.ErrNoRows) to prevent
+// context cancellation from being misclassified as a domain not-found condition
+// (S15 ctx-cancel classification fix).
+//
 // op is the method name used only in InternalMessage for operator debugging;
 // key is the lookup key surfaced the same way.
-func scanConfigOrMapError(row Row, op, key string) (*domain.ConfigEntry, []byte, *string, []byte, []byte, error) {
+func scanConfigOrMapError(ctx context.Context, row Row, op, key string) (*domain.ConfigEntry, []byte, *string, []byte, []byte, error) {
 	e, ct, keyID, edk, nonce, err := scanConfigRow(row)
 	if err == nil {
 		return e, ct, keyID, edk, nonce, nil
+	}
+	if infraErr := ctxCanceledInfraError(ctx, err, op, key); infraErr != nil {
+		return nil, nil, nil, nil, nil, infraErr
 	}
 	if errors.Is(err, pgx.ErrNoRows) {
 		return nil, nil, nil, nil, nil, &errcode.Error{
@@ -282,6 +301,7 @@ func scanConfigOrMapError(row Row, op, key string) (*domain.ConfigEntry, []byte,
 			Message:         "config not found",
 			InternalMessage: fmt.Sprintf("config repo: %s miss key=%s", op, key),
 			Cause:           err,
+			Category:        errcode.CategoryDomain,
 		}
 	}
 	return nil, nil, nil, nil, nil, &errcode.Error{
@@ -289,6 +309,7 @@ func scanConfigOrMapError(row Row, op, key string) (*domain.ConfigEntry, []byte,
 		Message:         "config repo query failed",
 		InternalMessage: fmt.Sprintf("config repo: %s scan error key=%s", op, key),
 		Cause:           err,
+		Category:        errcode.CategoryInfra,
 	}
 }
 
@@ -328,7 +349,7 @@ func (r *ConfigRepository) decryptScannedEntry(ctx context.Context, e *domain.Co
 func (r *ConfigRepository) GetByKey(ctx context.Context, key string) (*domain.ConfigEntry, error) {
 	const q = `SELECT ` + configEntryColumns + ` FROM config_entries WHERE key = $1`
 	row := r.resolveDB(ctx).QueryRow(ctx, q, key)
-	e, ct, keyID, edk, nonce, err := scanConfigOrMapError(row, "GetByKey", key)
+	e, ct, keyID, edk, nonce, err := scanConfigOrMapError(ctx, row, "GetByKey", key)
 	if err != nil {
 		return nil, err
 	}
@@ -383,12 +404,16 @@ func (r *ConfigRepository) Update(ctx context.Context, key string, value string)
 	var sensitive bool
 	selectRow := db.QueryRow(ctx, selectQ, key)
 	if scanErr := selectRow.Scan(&sensitive); scanErr != nil {
+		if infraErr := ctxCanceledInfraError(ctx, scanErr, "Update", key); infraErr != nil {
+			return nil, infraErr
+		}
 		if errors.Is(scanErr, pgx.ErrNoRows) {
 			return nil, &errcode.Error{
 				Code:            errcode.ErrConfigRepoNotFound,
 				Message:         "config not found",
 				InternalMessage: fmt.Sprintf("config repo: Update miss key=%s", key),
 				Cause:           scanErr,
+				Category:        errcode.CategoryDomain,
 			}
 		}
 		return nil, &errcode.Error{
@@ -396,6 +421,7 @@ func (r *ConfigRepository) Update(ctx context.Context, key string, value string)
 			Message:         "config repo query failed",
 			InternalMessage: fmt.Sprintf("config repo: Update select-for-update error key=%s", key),
 			Cause:           scanErr,
+			Category:        errcode.CategoryInfra,
 		}
 	}
 
@@ -440,7 +466,7 @@ func (r *ConfigRepository) doUpdate(ctx context.Context, db DBTX, key string, va
 		row = db.QueryRow(ctx, q, value, key)
 	}
 
-	e, ct, keyID, edk, nonce, scanErr := scanConfigOrMapError(row, "Update", key)
+	e, ct, keyID, edk, nonce, scanErr := scanConfigOrMapError(ctx, row, "Update", key)
 	if scanErr != nil {
 		return nil, scanErr
 	}
@@ -461,7 +487,7 @@ func (r *ConfigRepository) Delete(ctx context.Context, key string) (*domain.Conf
 		return nil, err
 	}
 	row := db.QueryRow(ctx, q, key)
-	e, ct, keyID, edk, nonce, scanErr := scanConfigOrMapError(row, "Delete", key)
+	e, ct, keyID, edk, nonce, scanErr := scanConfigOrMapError(ctx, row, "Delete", key)
 	if scanErr != nil {
 		return nil, scanErr
 	}
@@ -526,13 +552,13 @@ func (r *ConfigRepository) List(ctx context.Context, params query.ListParams) ([
 	b.Append("SELECT " + listEntryColumns + " FROM config_entries WHERE 1=1")
 
 	if err := query.AppendKeyset(b, params); err != nil {
-		return nil, errcode.Wrap(errcode.ErrConfigRepoQuery, "config repo: keyset build failed", err)
+		return nil, errcode.WrapInfra(errcode.ErrConfigRepoQuery, "config repo: keyset build failed", err)
 	}
 
 	sql, args := b.Build()
 	rows, err := r.resolveDB(ctx).Query(ctx, sql, args...)
 	if err != nil {
-		return nil, errcode.Wrap(errcode.ErrConfigRepoQuery, "config repo: list failed", err)
+		return nil, errcode.WrapInfra(errcode.ErrConfigRepoQuery, "config repo: list failed", err)
 	}
 	defer rows.Close()
 
@@ -545,7 +571,7 @@ func (r *ConfigRepository) List(ctx context.Context, params query.ListParams) ([
 			valueKeyID *string
 		)
 		if err := rows.Scan(&e.ID, &e.Key, &e.Value, &e.Sensitive, &e.Version, &e.CreatedAt, &e.UpdatedAt, &valueKeyID); err != nil {
-			return nil, errcode.Wrap(errcode.ErrConfigRepoQuery, "config repo: scan failed", err)
+			return nil, errcode.WrapInfra(errcode.ErrConfigRepoQuery, "config repo: scan failed", err)
 		}
 		if e.Sensitive {
 			r.applySensitiveListSentinel(ctx, &e, valueKeyID, currentID, staleLogged)
@@ -553,7 +579,7 @@ func (r *ConfigRepository) List(ctx context.Context, params query.ListParams) ([
 		entries = append(entries, &e)
 	}
 	if err := rows.Err(); err != nil {
-		return nil, errcode.Wrap(errcode.ErrConfigRepoQuery, "config repo: rows error", err)
+		return nil, errcode.WrapInfra(errcode.ErrConfigRepoQuery, "config repo: rows error", err)
 	}
 
 	return entries, nil
@@ -591,7 +617,7 @@ func (r *ConfigRepository) PublishVersion(ctx context.Context, version *domain.C
 		)
 	}
 	if err != nil {
-		return errcode.Wrap(errcode.ErrConfigRepoQuery,
+		return errcode.WrapInfra(errcode.ErrConfigRepoQuery,
 			fmt.Sprintf("config repo: publish version failed for config %s v%d",
 				version.ConfigID, version.Version), err)
 	}
@@ -617,12 +643,16 @@ func (r *ConfigRepository) GetVersion(ctx context.Context, configID string, vers
 		&v.ID, &v.ConfigID, &v.Version, &v.Value, &v.Sensitive, &v.PublishedAt,
 		&valueCipher, &valueKeyID, &valueEDK, &valueNonce,
 	); err != nil {
+		if infraErr := ctxCanceledInfraError(ctx, err, "GetVersion", configID); infraErr != nil {
+			return nil, infraErr
+		}
 		if errors.Is(err, pgx.ErrNoRows) {
 			return nil, &errcode.Error{
 				Code:            errcode.ErrConfigRepoNotFound,
 				Message:         "config version not found",
 				InternalMessage: fmt.Sprintf("config repo: GetVersion miss config_id=%s version=%d", configID, version),
 				Cause:           err,
+				Category:        errcode.CategoryDomain,
 			}
 		}
 		return nil, &errcode.Error{
@@ -630,6 +660,7 @@ func (r *ConfigRepository) GetVersion(ctx context.Context, configID string, vers
 			Message:         "config repo query failed",
 			InternalMessage: fmt.Sprintf("config repo: GetVersion scan error config_id=%s version=%d", configID, version),
 			Cause:           err,
+			Category:        errcode.CategoryInfra,
 		}
 	}
