+++ cells/configcore/internal/adapters/postgres/config_repo_test.go
@@ -105,6 +105,68 @@ func TestGetByKey_OtherScanError_ReturnsErrConfigRepoQuery(t *testing.T) {
 	assert.Equal(t, errcode.ErrConfigRepoQuery, ec.Code)
 }
 
+// TestGetByKey_NotFound_HasDomainCategory locks ErrConfigRepoNotFound's Category
+// to CategoryDomain so that errcode.IsDomainNotFound dual-channel check works
+// (S15 Category fix).
+func TestGetByKey_NotFound_HasDomainCategory(t *testing.T) {
+	db := &mockDB{queryRowResult: &mockRow{scanErr: pgx.ErrNoRows}}
+	repo := newConfigRepositoryFromDBTX(db)
+
+	_, err := repo.GetByKey(context.Background(), "missing")
+	require.True(t, errcode.IsDomainNotFound(err, errcode.ErrConfigRepoNotFound),
+		"ErrConfigRepoNotFound must have Category=CategoryDomain for IsDomainNotFound to work")
+	require.False(t, errcode.IsInfraError(err),
+		"domain not-found must not be treated as infra")
+}
+
+// TestConfigRepo_CtxCanceled_ClassifiedAsInfra verifies that context.Canceled
+// and context.DeadlineExceeded are always classified as infra (never domain
+// not-found), across GetByKey, Update SELECT-FOR-UPDATE, and GetVersion
+// (S15 ctx-cancel classification fix).
+func TestConfigRepo_CtxCanceled_ClassifiedAsInfra(t *testing.T) {
+	tests := []struct {
+		name    string
+		scanErr error
+	}{
+		{"ctx canceled", context.Canceled},
+		{"ctx deadline exceeded", context.DeadlineExceeded},
+	}
+	for _, tc := range tests {
+		tc := tc
+		t.Run("GetByKey/"+tc.name, func(t *testing.T) {
+			db := &mockDB{queryRowResult: &mockRow{scanErr: tc.scanErr}}
+			repo := newConfigRepositoryFromDBTX(db)
+
+			_, err := repo.GetByKey(context.Background(), "any")
+			require.Error(t, err)
+			require.True(t, errcode.IsInfraError(err))
+			require.False(t, errcode.IsDomainNotFound(err, errcode.ErrConfigRepoNotFound),
+				"ctx cancel must not leak into domain not-found branch")
+			var ec *errcode.Error
+			require.ErrorAs(t, err, &ec)
+			assert.Equal(t, errcode.ErrConfigRepoQuery, ec.Code)
+		})
+		t.Run("Update_SelectForUpdate/"+tc.name, func(t *testing.T) {
+			seqDB := &sequencedMockDB{rows: []*mockRow{{scanErr: tc.scanErr}}}
+			repo := newConfigRepositoryFromDBTX(seqDB)
+
+			_, err := repo.Update(context.Background(), "k", "v")
+			require.Error(t, err)
+			require.True(t, errcode.IsInfraError(err))
+			require.False(t, errcode.IsDomainNotFound(err, errcode.ErrConfigRepoNotFound))
+		})
+		t.Run("GetVersion/"+tc.name, func(t *testing.T) {
+			db := &mockDB{queryRowResult: &mockRow{scanErr: tc.scanErr}}
+			repo := newConfigRepositoryFromDBTX(db)
+
+			_, err := repo.GetVersion(context.Background(), "cfg-1", 1)
+			require.Error(t, err)
+			require.True(t, errcode.IsInfraError(err))
+			require.False(t, errcode.IsDomainNotFound(err, errcode.ErrConfigRepoNotFound))
+		})
+	}
+}
+
 // TestConfigRepository_GetByKey_NotFound is a legacy name kept for backward
 // reference. It tests the other-error path (assert.AnError != pgx.ErrNoRows).
 func TestConfigRepository_GetByKey_NotFound(t *testing.T) {
@@ -160,6 +222,8 @@ func TestConfigRepository_Update_NotFound(t *testing.T) {
 	var ec *errcode.Error
 	require.ErrorAs(t, err, &ec)
 	assert.Equal(t, errcode.ErrConfigRepoNotFound, ec.Code)
+	require.True(t, errcode.IsDomainNotFound(err, errcode.ErrConfigRepoNotFound),
+		"Update not-found must have Category=CategoryDomain")
 }
 
 // TestConfigRepository_UpdateForRollback tests the 4-arg UpdateForRollback method.
@@ -196,6 +260,8 @@ func TestConfigRepository_UpdateForRollback_NotFound(t *testing.T) {
 	var ec *errcode.Error
 	require.ErrorAs(t, err, &ec)
 	assert.Equal(t, errcode.ErrConfigRepoNotFound, ec.Code)
+	require.True(t, errcode.IsDomainNotFound(err, errcode.ErrConfigRepoNotFound),
+		"UpdateForRollback not-found must have Category=CategoryDomain")
 }
 
 func TestConfigRepository_Delete(t *testing.T) {
@@ -229,6 +295,8 @@ func TestConfigRepository_Delete_NotFound(t *testing.T) {
 	var ec *errcode.Error
 	require.ErrorAs(t, err, &ec)
 	assert.Equal(t, errcode.ErrConfigRepoNotFound, ec.Code)
+	require.True(t, errcode.IsDomainNotFound(err, errcode.ErrConfigRepoNotFound),
+		"Delete not-found must have Category=CategoryDomain")
 }
 
 func TestConfigRepository_List(t *testing.T) {
@@ -416,6 +484,8 @@ func TestGetVersion_NotFound_ReturnsErrConfigRepoNotFound(t *testing.T) {
 	var ec *errcode.Error
 	require.ErrorAs(t, err, &ec)
 	assert.Equal(t, errcode.ErrConfigRepoNotFound, ec.Code)
+	require.True(t, errcode.IsDomainNotFound(err, errcode.ErrConfigRepoNotFound),
+		"GetVersion not-found must have Category=CategoryDomain")
 }
 
 // TestGetVersion_OtherScanError_ReturnsErrConfigRepoQuery verifies that scan
